chore: 폴더 위치 변경
