+++ src/app/[locale]/(auth)/signup/page.tsx
@@ -1,17 +1,23 @@
-import { getTranslations } from 'next-intl/server';
-import { Locale } from '@/i18n';
+'use client';
 
-interface SignUpPageProps {
-  params: Promise<{ locale: Locale }>;
-}
+import { useSignupImageSize } from '@/features/signup/hooks/useSignupImageSize';
+import SignupForm from '@/features/signup/ui/SignupForm';
+import { LoginImageIcon } from '@/shared/ui/icon';
 
-export default async function SignUpPage({ params }: SignUpPageProps) {
-  const { locale } = await params;
-  const t = await getTranslations({ locale, namespace: 'pages.signup' });
+export default function SignUpPage() {
+  const size = useSignupImageSize();
 
   return (
-    <div className="flex flex-1 items-center justify-center">
-      <div className="text-2xl font-bold">{t('title')}</div>
+    <div className="max-w-web web:flex-row flex flex-1 flex-col items-center justify-between">
+      <div className="flex flex-col items-center text-center">
+        <h2 className="text-2xl font-semibold text-gray-800">Welcome to 같이 달램!</h2>
+        <p className="text-base">
+          바쁜 일상 속 잠깐의 휴식, <br />
+          이제는 같이 달램과 함께 해보세요
+        </p>
+        <LoginImageIcon size={size} />
+      </div>
+      <SignupForm />
     </div>
   );
 }
