+++ src/features/signup/utils/getSignupImageSize.ts
