+++ src/shared/ui/InputWithLabel/index.ts
@@ -1,0 +1,1 @@
+export { InputWithLabel } from './InputWithLabel';
